Enable PragmataPro Fraktur ligatures
I initially had them disabled because I was having trouble with the s's.
I kept mistaking them for f's. Words like "stressful" involve some extra
parsing time because it looks like "ftreffful" with ligatures. But at
the same time, it's kind of growing on me and I kind of like it.