Hoist configuration flake outputs into root flake
I want to define a flakeref with a relative path outside of the flake
directory. Turns out that you can't do this because the referenced flake
doesn't get copied into the Nix store during evaluation. So I'm hoisting
the configuration flake outputs into the root directory to preserve the
hierarchical structure of flakes and subflakes references.