Expose packages in flake outputs
This commit exposes the packages that are included in the devShell
environment as an explicit flake output for convenience. This will allow
the root "motherflake" to pass these inputs into configurations as
special args which can be used to derive the list of packages to include
in the system profile.