Create `TooltipProvider` component
This component is responsible for managing the tooltip state across all
instances of the `Tooltip` component. It uses a nanostore to keep track
of the active tooltip and whether it is currently visible.

The `TooltipProvider` component should be mounted at the document root
to ensure that the tooltip is rendered at the top-most stacking context
and can be positioned relative to the viewport (and so that we can set
a z-index value that will ensure the tooltip is always on top).