Position gradients with CSS instead of GSAP
When setting the positions inside of a scroll event handler, the
positions lag behind the actual scroll position. This is noticable when
scrolling quickly.

Instead, we can use absolute positioning to anchor the gradients to the
top and bottom of the container. This involved rearranging the DOM
structure so that the gradients exist outside of the scroll container.