Split `ScrollContainer` into multiple components
This commit splits the ScrollContainer component into static and dynamic
components. Before, the static content rendered within a ScrollContainer
component could not be selectively hydrated in an Astro island because
the gradient container (with the client:only directive applied) wrapped
the entire content, skipping server-side rendering for all of the static
content.

Now, the ScrollContainer component is split into three components:

✦ ScrollContainer: The static content that should be rendered on the
  server (selective hydration).

✦ ScrollGradientContainer: The container that wraps the ScrollContainer
  and ScrollGradient components. Mainly used to apply styles to the
  gradient overlay with relative > absolute positioning.

✦ ScrollGradient: The dynamic gradient overlay that fades in and out
  based on the scroll position that should be rendered on the client.