Write note durations and bar lines to output
Adds support for writing note durations and explicit bar lines to the
LilyPond output. If a note is longer than a beat, it is split into
multiple beats so as not to obscure the beat structure.

Does not yet support tuplets or other complex note durations. Only the
most common note durations are supported (and their dotted variants).