Refactor SCSS to remove duplicated styles
I noticed that global styles were being duplicated for each @use rule
that existed in the CSS modules. It turns out that SCSS and CSS modules
operate at different stages of the styling process.

SCSS is a preprocessor for CSS. It runs before the CSS is compiled. CSS
modules operate during a separate step at build time. While @use is
intended to handle duplication, it doesn't work as expected when used
with CSS modules.

✦ Separated variables and mixins from global rule sets.
✦ Import variables and mixins globally in Vite.
✦ Import global rule sets in entry point.
✦ Removed @use rules from CSS modules.